refactor: move IconButton component to the separate folder
LP-4
