+++ src/components/IconButton/IconButton.module.scss
@@ -1,0 +1,5 @@
+.iconButton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
