+++ src/interfaces/Svg.ts
@@ -1,6 +1,5 @@
 export interface Svg {
   id: string;
-  blockName?: string;
   width: string;
   height: string;
   fill?: string;
